improvement 1 done

+++ SA-PROJET-LIGHT-ONE-Q/src/algorithms/quantum_search_algorithm.rs
@@ -7,7 +7,7 @@ use rand::Rng;
 use std::collections::HashMap;
 
 // function run quantum search algorithm 10 times and return accuracy of results
-pub fn quantum_search_algorithm_wrapper(dataset: &Vec<i32>, target: i32, iterations: usize) -> f64 {
+pub fn quantum_search_algorithm_wrapper(dataset: &Vec<i32>, target: i32, iterations: u32) -> f64 {
     let mut total_accuracy = 0.0;
  
     for _ in 0..10 {
@@ -20,7 +20,7 @@ pub fn quantum_search_algorithm_wrapper(dataset: &Vec<i32>, target: i32, iterati
     total_accuracy / 10.0
 }
 
-pub fn quantum_search(dataset: &Vec<i32>, target: i32, iterations: usize) -> Option<i32> {
+pub fn quantum_search(dataset: &Vec<i32>, target: i32, iterations: u32) -> Option<i32> {
     let mut rng = rand::thread_rng(); // Random number generator
     // Step 0: Initialize probabilities
     let mut probabilities: HashMap<usize, f64> = (0..dataset.len())
